requirements.txt
# ----- Core deep-learning stack -----
tensorflow==2.15.0          # includes Keras 2.15 and EfficientNetB0
numpy==1.26.4               # upper range supported by TF-2.15

# ----- Classical ML / metrics -----
scikit-learn==1.3.2         # works with numpy 1.26+

# ----- Image I/O -----
Pillow==10.2.0              # provides the PIL module

# ----- REST API -----
Flask==2.3.3
Werkzeug==2.3.7             # Flask’s WSGI/utility library (explicit for reproducibility)

# ----- (optional) production WSGI server -----
# gunicorn==21.2.0